api/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
beautifulsoup4
# Add other dependencies from original requirements.txt if any, but modules seemed to use standard libs or these.
